Oversized commit of transformation code Why I couldn't split the code into chunks for separate commits so it all has to go in one large and broken commit.
This introduces the code for transforming the data in Pandas. I have to use
Pandas because Google Sheets blocked Heroku and I can't use the Herkou
postgres database as a result. I may move to AWS in the future once I've
developed a front end for the responses with its own database.

How
- Remove OOP. It's unnecessary here
- Add logic to the sheet pull function to set the first row as the header
- Add a function to clean up the initial dataframe
- Add a function to mark the 5 days before the first instance of an event group
as belonging to that group
- Add a function to create dataframes to be used by the algorithms. THIS IS
STILL BROKEN

Note
I need to get rid of the test functions at the bottom