Reconnect the dataframe to the prediction Why This function will connect the list of predictions back to the appropriate item in the dataframe. This will be useful for visualization and further analysis in a database.
How
- Add a dataframe to the prep data function that preserves the timestamp
and the data in string format
- Add the reconnect function which attaches the list of predictions as
a column on the dataframe and converts them to boolean

Note
The other "boolean" columns are actually stored as strings. I was previously
aware of this problem and ignored it. I may return and fix it if it becomes
more of an issue.