Add Flask app for recording survey answers
Why
This is the initial version of the app that
will be used to record answers on a mobile friendly website.
It replaces the Google Form set-up

How
Create a Flask app that has temp storage for questions and
allows users to submit answers. Note that this is still a work
in progress and may not work in its current state.
Note that this is still a work in progress